Updated Home, Login and signup. Bugfix and improvements
Home:
check if the user is logged in, if is not, the signin button is displayed, user profile otherwise.

Login:
same as home, deleted signup button (if you are trying to log in why you want to signup?)

Signup:
same as login. Added passwordCheck.js to check is password and passwordcheck forms contains the same
value

index.jsp:
added foreach statments, now the page is dynamic